requirements.txt
azure-storage-blob
metpx-sarracenia
